+++ client/src/App.tsx
@@ -1,0 +1,5 @@
+function App() {
+  return <div className="App">Learn React</div>;
+}
+
+export default App;
