--- client/src/components/Header/assets/index.js
@@ -1,1 +1,0 @@
-export { default as logo } from './logo.svg';
