--- client/src/services/index.js
@@ -1,15 +1,0 @@
-import todoLists from '../data/todos';
-
-const service = {
-  getData: ({ from, to }) => {
-    return new Promise((resolve) => {
-      const data = todoLists.slice(from, to);
-      resolve({
-        count: todoLists.length,
-        data: data,
-      });
-    });
-  },
-};
-
-export default service;
