--- client/src/views/CreateTodoListPage/CreateTodoListPage.js
@@ -1,3 +1,0 @@
-export default function CreateTodoListPage() {
-  return <div>CreateTodoListPage</div>;
-}
