--- client/src/views/NotFoundPage/NotFoundPage.js
@@ -1,3 +1,0 @@
-export default function NotFoundPage() {
-  return <div>NotFoundPage</div>;
-}
